update to 16th note and swing

+++ display.h
@@ -1,5 +1,4 @@
 #pragma once
 
 void setup_display ( );
-void display_bpm (float);
-
+void display_bpm (int);
